refactor: internalize jsonrpc implementation
