+++ src/main/java/com/amannmalik/mcp/api/CreateMessageRequest.java
@@ -1,5 +1,6 @@
 package com.amannmalik.mcp.api;
 
+import com.amannmalik.mcp.jsonrpc.JsonCodec;
 import com.amannmalik.mcp.core.AbstractEntityCodec;
 import com.amannmalik.mcp.sampling.ModelPreferences;
 import com.amannmalik.mcp.util.ValidationUtil;
