+++ src/main/java/com/amannmalik/mcp/api/ProgressNotification.java
@@ -1,5 +1,6 @@
 package com.amannmalik.mcp.api;
 
+import com.amannmalik.mcp.jsonrpc.JsonCodec;
 import com.amannmalik.mcp.util.ValidationUtil;
 import jakarta.json.*;
 
@@ -38,6 +39,10 @@ public record ProgressNotification(
             return new ProgressNotification(token, progress, total, message);
         }
     };
+
+    public static JsonObject toJson(ProgressNotification note) {
+        return CODEC.toJson(note);
+    }
 
     public static Optional<ProgressToken> fromMeta(JsonObject params) {
         if (params == null || !params.containsKey("_meta")) return Optional.empty();
