+++ src/main/java/com/amannmalik/mcp/api/SamplingMessage.java
@@ -1,5 +1,6 @@
 package com.amannmalik.mcp.api;
 
+import com.amannmalik.mcp.jsonrpc.JsonCodec;
 import com.amannmalik.mcp.sampling.MessageContent;
 import com.amannmalik.mcp.core.AbstractEntityCodec;
 import jakarta.json.Json;
