+++ src/main/java/com/amannmalik/mcp/core/ResourceMetadata.java
@@ -1,6 +1,6 @@
 package com.amannmalik.mcp.core;
 
-import com.amannmalik.mcp.api.JsonCodec;
+import com.amannmalik.mcp.jsonrpc.JsonCodec;
 import jakarta.json.*;
 
 import java.util.List;
