+++ src/main/java/com/amannmalik/mcp/logging/SetLevelRequest.java
@@ -2,7 +2,7 @@ package com.amannmalik.mcp.logging;
 
 import com.amannmalik.mcp.api.LoggingLevel;
 import com.amannmalik.mcp.core.AbstractEntityCodec;
-import com.amannmalik.mcp.api.JsonCodec;
+import com.amannmalik.mcp.jsonrpc.JsonCodec;
 import com.amannmalik.mcp.util.ValidationUtil;
 import jakarta.json.*;
