+++ src/main/java/com/amannmalik/mcp/prompts/PromptListChangedNotification.java
@@ -1,7 +1,7 @@
 package com.amannmalik.mcp.prompts;
 
 import com.amannmalik.mcp.core.AbstractEntityCodec;
-import com.amannmalik.mcp.api.JsonCodec;
+import com.amannmalik.mcp.jsonrpc.JsonCodec;
 
 public record PromptListChangedNotification() {
     public static final JsonCodec<PromptListChangedNotification> CODEC =
