+++ src/main/java/com/amannmalik/mcp/resources/ReadResourceResult.java
@@ -1,6 +1,6 @@
 package com.amannmalik.mcp.resources;
 
-import com.amannmalik.mcp.api.JsonCodec;
+import com.amannmalik.mcp.jsonrpc.JsonCodec;
 import com.amannmalik.mcp.api.ResourceBlock;
 import com.amannmalik.mcp.util.Immutable;
 import com.amannmalik.mcp.util.ValidationUtil;
