+++ src/main/java/com/amannmalik/mcp/resources/ResourceListChangedNotification.java
@@ -1,7 +1,7 @@
 package com.amannmalik.mcp.resources;
 
 import com.amannmalik.mcp.core.AbstractEntityCodec;
-import com.amannmalik.mcp.api.JsonCodec;
+import com.amannmalik.mcp.jsonrpc.JsonCodec;
 
 public record ResourceListChangedNotification() {
     public static final JsonCodec<ResourceListChangedNotification> CODEC =
