+++ src/main/java/com/amannmalik/mcp/roots/RootsListChangedNotification.java
@@ -1,7 +1,7 @@
 package com.amannmalik.mcp.roots;
 
 import com.amannmalik.mcp.core.AbstractEntityCodec;
-import com.amannmalik.mcp.api.JsonCodec;
+import com.amannmalik.mcp.jsonrpc.JsonCodec;
 
 public record RootsListChangedNotification() {
     public static final JsonCodec<RootsListChangedNotification> CODEC =
