+++ src/main/java/com/amannmalik/mcp/sampling/ModelHint.java
@@ -1,7 +1,7 @@
 package com.amannmalik.mcp.sampling;
 
 import com.amannmalik.mcp.core.AbstractEntityCodec;
-import com.amannmalik.mcp.api.JsonCodec;
+import com.amannmalik.mcp.jsonrpc.JsonCodec;
 import com.amannmalik.mcp.util.ValidationUtil;
 import jakarta.json.*;
