+++ src/main/java/com/amannmalik/mcp/tools/ToolAnnotations.java
@@ -1,7 +1,7 @@
 package com.amannmalik.mcp.tools;
 
 import com.amannmalik.mcp.core.AbstractEntityCodec;
-import com.amannmalik.mcp.api.JsonCodec;
+import com.amannmalik.mcp.jsonrpc.JsonCodec;
 import com.amannmalik.mcp.util.ValidationUtil;
 import jakarta.json.*;
