+++ src/main/java/com/amannmalik/mcp/util/CancelledNotification.java
@@ -1,6 +1,6 @@
 package com.amannmalik.mcp.util;
 
-import com.amannmalik.mcp.api.JsonCodec;
+import com.amannmalik.mcp.jsonrpc.JsonCodec;
 import com.amannmalik.mcp.jsonrpc.RequestId;
 import jakarta.json.*;
